requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
defusedxml==0.7.1
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
httpie==3.2.4
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
Pygments==2.18.0
PyJWT==2.10.0
PySocks==1.7.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
setuptools==75.5.0
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
